+++ src/test/java/carracing/domain/PositionRestrictionsStrategy.java
