+++ src/test/java/carracing/domain/RandomMaxValue.java
