+++ src/test/java/carracing/domain/RandomMinValue.java
